update pysdk and format code
