+++ sdks/pysdk/base.py
@@ -182,8 +182,3 @@ class TransferEvtAbi(BaseType):
         # since 'from' is a keyword of python
         args = {'from': _from, 'to': to, 'amount': amount}
         super().__init__(**args)
-
-
-class AddMetaAbi(BaseType):
-    def __init__(self, key, value, creator):
-        super().__init__(key=key, value=value, creator=creator)
